refactor: AuthButton 컴포넌트 분리 및 로그인 로직 유틸화
- Header 내부 로그인/로그아웃 조건부 렌더링 분리 → AuthButton 컴포넌트 생성
- GitHub OAuth 리디렉션 로직을 redirectToGithubAuthorize 유틸 함수로 분리
